fix: global install must reinstall the whole package
includes dependencies
