+++ lib/bin.js
@@ -12,6 +12,7 @@
  * Module dependencies.
  */
 
+const chalk = require('chalk');
 const debug = require('debug')('npminstall:bin');
 const path = require('path');
 const fs = require('mz/fs');
@@ -31,9 +32,14 @@ function* bin(parentDir, pkg, pkgDir, options) {
   if (names.length === 0) return;
 
   // root package link to options.binDir first
-  const binDir = options.binDir && parentDir === options.targetDir
-    ? options.binDir
-    : path.join(parentDir, 'node_modules/.bin');
+  let binDir;
+  let showBinLog = false;
+  if (options.binDir && parentDir === options.targetDir) {
+    binDir = options.binDir;
+    showBinLog = true;
+  } else {
+    binDir = path.join(parentDir, 'node_modules', '.bin');
+  }
   yield utils.mkdirp(binDir);
 
   for (const name of names) {
@@ -41,8 +47,13 @@ function* bin(parentDir, pkg, pkgDir, options) {
     const destBin = path.join(binDir, name);
     yield fs.chmod(srcBin, 0o755);
     const relative = yield linkBin(srcBin, destBin);
-    debug('[%s@%s] link %s => %s',
-      pkg.name, pkg.version, srcBin, relative);
+    if (showBinLog) {
+      options.console.info('[%s] link %s@ -> %s',
+        chalk.green(pkg.name + '@' + pkg.version), chalk.magenta(destBin), relative);
+    } else {
+      debug('[%s@%s] link %s@ -> %s',
+        pkg.name, pkg.version, destBin, relative);
+    }
   }
 }
