+++ lib/install.js
@@ -167,6 +167,7 @@ function* _install(parentDir, pkg, options) {
     });
   }
   debug('[%s@%s] installed', realPkg.name, realPkg.version);
+
   return {
     exists: false,
     dir: realPkgDir,
