+++ lib/utils.js
@@ -196,7 +196,7 @@ exports.copyInstall = function* (src, options) {
   }
 
   const targetdir = exports.getPackageStorePath(options.storeDir, realPkg);
-  const donefile = path.join(targetdir, '.tnpminstall.done');
+  const donefile = path.join(targetdir, '.npminstall.done');
   const key = `copy:${targetdir}`;
   const result = {
     dir: targetdir,
